+++ packages/core/styles/src/packages/atomic/fileUploader/index.ts
@@ -3,11 +3,13 @@ import {
   fileUploaderSprinkle,
   backgroundColor,
   borderColor,
+  overlayBorderRadius,
 } from "./nimbus-fileUploader.css";
 
 export const fileUploaderVars = {
   backgroundColor,
   borderColor,
+  overlayBorderRadius,
 };
 
 export const fileUploader = {