+++ packages/core/styles/CHANGELOG.md
@@ -8,6 +8,12 @@ Nimbus Styles deprive all styles needed to build nimbus components.
 
 - `List`: Added `padding` property to set the padding of the list. ([#390](https://github.com/TiendaNube/nimbus-design-system/pull/390) by [@joacotornello](https://github.com/joacotornello))
 
+## 2025-12-11 `9.44.1`
+
+#### 🐛 Bug fixes
+
+- Fixes `FileUploader` overlay styles to use the correct background color and border color. ([#388](https://github.com/TiendaNube/nimbus-design-system/pull/388) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-12-09 `9.44.0`
 
 #### 🎉 New features