+++ packages/react/src/atomic/FileUploader/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The File Uploader component is used to upload files.
 
+## 2025-12-11 `2.8.1`
+
+#### 🐛 Bug fixes
+
+- Fixes `FileUploaderOverlay` component to use the correct background color and border color. ([#388](https://github.com/TiendaNube/nimbus-design-system/pull/388) by [@joacotornello](https://github.com/joacotornello))
+- Fixes `FileUploader` to use `asOverlay` property to render a custom overlay when dragging files over the uploader. Better API usage. ([#388](https://github.com/TiendaNube/nimbus-design-system/pull/388) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-12-09 `2.8.0`
 
 #### 🎉 New features